requirements.txt
streamlit
tensorflow==2.13.0
opencv-python
numpy
Pillow
typing-extensions<4.6.0

